- Removed own rounding code in favor of boost's rounding code. - Added in Vector2 type converters, all non floating point types have their values rounded and type converted.

+++ simplerpg/Map.h
@@ -39,7 +39,7 @@ public:
 		return mMapData[x][y].tile; 
 	}
 
-	vector<Vector2f> *search(Vector2f start, Vector2f end);
+	vector<Vector2f> *search(Vector2i start, Vector2i end);
 
 	void renderMap(Rect rect, WINDOW *wnd);
